VILE PRIVATE/requirements.txt
aioconsole==0.5.1
aiocron==1.8
aiodns==3.0.0
aiofiles==22.1.0
aiohttp==3.8.3
aiohttp-cors==0.7.0
aiohttp-requests==0.1.3
aiomysql==0.1.1
aioredis==2.0.1
aiosignal==1.3.1
aiosqlite==0.17.0
anyio==3.6.2
appdirs==1.4.4
argcomplete==2.0.0
arrow==1.2.3
asgiref==3.6.0
astunparse==1.6.3
async-cse==0.3.0
async-generator==1.10
async-pytube==1.0.2
async-timeout==4.0.2
asynccpu==1.2.2
asyncer==0.0.2
asyncio==3.4.3
asyncio-extras==1.3.2
asyncmy==0.2.5
asyncpg==0.27.0
asyncpraw==7.3.0
asyncprawcore==2.2.0
asyncstdlib==3.10.5
attrs==22.1.0
Automat==0.8.0
beautifulsoup4==4.11.1
bella==2.1.5
bernhard==0.2.6
better-ipc==1.1.4
betterstatcord.py==3.4.1
black==22.10.0
bleach==3.3.0
blinker==1.4
blobfile==2.0.1
boltons==23.0.0
bot-studio==1.4.0
bottle==0.12.23
braceexpand==0.1.7
Brotli==1.0.9
bs4==0.0.1
button-paginator @ git+https://github.com/cop-discord/button_paginator.git@51199d2c46749b21b3c5f57556a504e58cac3413
cairocffi==1.4.0
CairoSVG==2.5.2
cassandra-driver==3.25.0
cchardet==2.1.7
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.0
chardet==3.0.4
charset-normalizer==2.1.1
chess==1.9.2
chevron==0.14.0
click==8.1.3
colorama==0.4.4
coloredlogs==14.3
colorgram.py==1.2.0
colorthief==0.2.1
configobj==5.0.6
constantly==15.1.0
contourpy==1.0.6
CouchDB==1.2
croniter==1.3.4
cryptography==2.8
cssselect==1.2.0
cssselect2==0.7.0
cycler==0.11.0
dataclass-factory==2.16
dataclasses-json==0.5.7
datamodel-code-generator==0.14.0
decorator==4.4.2
deep-translator==1.9.1
defusedxml==0.7.1
discord.py @ git+https://github.com/Z6g7mwQ35y4Bv/discord.py@3f248528f0068dd0f4dff1c387961eb5957af12e
distlib==0.3.6
distro==1.4.0
dnspython==2.2.1
docker==6.0.1
docopt==0.6.2
durations-nlp==1.0.1
elastic-transport==8.4.0
elasticsearch==8.4.3
email-validator==1.3.0
emcache==1.0.0
english-words==1.1.0
entrypoints==0.3
exceptiongroup==1.1.1
fake-useragent==1.1.1
fastapi==0.95.0
ffmpeg==1.4
filelock==3.9.0
filetype==1.2.0
fire==0.3.1
Flask==2.2.2
flatbuffers==23.3.3
fonttools==4.38.0
frozenlist==1.3.3
future==0.18.2
fuzzywuzzy==0.18.0
genson==1.2.2
geographiclib==2.0
geomet==0.2.1.post1
geopy==2.3.0
Glances==3.2.7
google-search-results==2.4.1
google-search-scraper-python==1.0.0
googletrans==3.0.0
graphitesender==0.11.2
greenlet==2.0.1
gTTS==2.3.0
h11==0.14.0
h2==3.2.0
h3==3.7.6
HastebinAPI==0.1.3
hpack==3.0.0
hstspreload==2022.9.1
html5lib==1.1
httpcore==0.16.3
httplib2==0.14.0
httpx==0.23.1
humanfriendly==10.0
humanize==4.4.0
hyperframe==5.2.0
hyperlink==19.0.0
identify==2.2.10
idna==2.8
ifaddr==0.2.0
ImageHash==4.3.1
imageio==2.25.0
imageio-ffmpeg==0.4.8
import-expression==1.1.4
importlib==1.0.4
importlib-metadata==4.6.4
importlib-resources==5.12.0
incremental==16.10.1
inflect==5.6.2
influxdb==5.3.1
influxdb-client==1.33.0
iniconfig==2.0.0
instagram-private-api==1.6.0.0
instaloader==4.9.6
isodate==0.6.1
isort==5.10.1
itsdangerous==2.1.2
javascript==1!1.0.1
jeepney==0.7.1
Jinja2==3.1.2
jishaku @ git+https://github.com/Z6g7mwQ35y4Bv/jishaku@a23de9cb8a49e8471227ab332f4ce29a61793b36
Js2Py==0.74
jsonm==1.1.1
jsonschema==3.2.0
kafka-python==2.0.2
kdtree==0.16
keyring==23.5.0
kiwisolver==1.4.4
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
legofy @ git+https://github.com/Z6g7mwQ35y4Bv/nsb.api.image.legofy@7e5daa58d047c253929b260b0816c84e8746db31
line-profiler==4.0.3
llvmlite==0.39.1
loguru==0.7.0
lxml==4.9.2
magickwand==0.2
MarkupSafe==2.1.1
marshmallow==3.16.0
marshmallow-enum==1.5.1
matplotlib==3.6.2
maturin==0.14.16
meson==0.53.2
more-itertools==8.10.0
motor==3.0.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.4
multidict==6.0.2
mutagen==1.46.0
mypy-extensions==0.4.3
nekos.py==1.0.3
nest-asyncio==1.5.5
netifaces==0.11.0
networkx==3.1
nodeenv==1.6.0
numba==0.56.4
numpy==1.23.5
oauthlib==3.1.1
onnxruntime==1.14.1
openai==0.26.4
openapi-schema-validator==0.1.6
openapi-spec-validator==0.3.3
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
orjson==3.8.3
outcome==1.2.0
package-name==0.1
packaging==21.3
paho-mqtt==1.6.1
parse==1.19.0
parsedatetime==2.6
parsel==1.7.0
pathspec==0.10.3
pbkdf2==1.3
pbr==5.11.1
pendulum==2.1.2
pexpect==4.6.0
pika==1.3.0
Pillow==9.3.0
pipreqs==0.4.11
platformdirs==2.6.2
playwright @ git+https://github.com/microsoft/playwright-python@56a40a5eecc1b8622f3696862d0981ade29cbb52
pluggy==1.0.0
ply==3.11
pooch==1.7.0
potsdb==1.0.3
powerline-shell==0.7.0
prance==0.22.11.4.0
pre-commit==2.13.0
proglog==0.1.10
prometheus-async==22.2.0
prometheus-client==0.14.1
protobuf==4.21.7
psutil==5.9.4
ptyprocess==0.7.0
py-spy==0.3.14
py3nvml==0.2.7
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycares==4.2.2
pycparser==2.21
pycryptodomex==3.17
pydantic==1.10.2
pydub==0.25.1
pyee==8.2.2
pygal==3.0.0
PyHamcrest==1.9.0
pyjsparser==2.7.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyMatting==1.1.8
pymdstat==0.4.2
pymongo==4.1.1
PyMySQL==1.0.2
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
pyppeteer==1.0.2
pyproject==1.3.1
pyquery==2.0.0
pyreadline==2.1
pyreadline3==3.4.1
pyrsistent==0.16.1
pysmi==0.3.4
pysnmp==4.4.12
PySnooper==1.1.1
PySocks==1.7.1
pytesseract==0.3.10
pytest==7.2.1
pytest-asyncio==0.20.3
python-dateutil==2.8.2
python-dotenv==0.15.0
python-multipart==0.0.6
python-youtube==0.6.4
pytube==12.1.0
pytube3==9.6.4
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
PyWavelets==1.4.1
PyYAML==5.3.1
pyzmq==24.0.1
random-user-agent==1.0.1
reactivex==4.0.4
redis==4.4.0
regex==2022.10.31
rembg==2.0.32
requests==2.28.1
requests-futures==1.0.0
requests-html==0.10.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
rfc3986==1.5.0
rivalapi @ git+https://github.com/cop-discord/rivalapi.git@5136cc07c58556e01f6353f00c7d3a05a3782b15
roblox==2.0.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scandir==1.10.0
scapy==2.5.0
schedule==1.1.0
scikit-image==0.19.3
scipy==1.10.1
SecretStorage==3.3.3
selenium==4.8.0
semver==2.13.0
service-identity==18.1.0
shazamio==0.3.1.1
simplejson==3.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==1.9.5
speedtest==0.0.1
speedtest-cli==2.1.2
ssh-import-id==5.10
starlette==0.26.1
starship==0.1.0
statcord.py==3.1.0
statsd==3.3.0
sympy==1.11.1
tabulate==0.9.0
termcolor==1.1.0
testresources==2.0.1
tifffile==2023.3.21
TikTokApi==5.2.2
tiktoken==0.1.2
timeago==1.0.16
timer==0.2.2
timezonefinder==6.1.9
tinycss2==1.1.1
tls-client==0.1.8
toml==0.10.2
tomli==2.0.1
topggpy==1.4.0
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
tuuid==1.0
tweepy==3.10.0
TwitterAPI==2.7.12
TwoCaptcha==0.0.1
textract==1.6.5
typed-ast==1.5.4
typing-inspect==0.7.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
ua-parser==0.10.0
ujson==4.3.0
undetected-chromedriver==3.2.1
Unidecode==1.3.6
update-checker==0.18.0
urbandict==0.6.1
urllib3==1.26.14
uvicorn==0.21.1
uvloop==0.17.0
validators==0.20.0
virtualenv==20.4.7
w3lib==2.1.1
wadllib==1.3.3
Wand==0.6.10
watchdog==3.0.0
watchfiles==0.18.1
webencodings==0.5.1
websocket-client==1.5.1
websocket-server==0.6.4
websockets==10.4
Werkzeug==2.2.2
wifi==0.3.8
wikipedia==1.4.0
winerp==1.4.0
wordcloud @ git+https://github.com/treyt3n/word_cloud@dbf7ab7753a36e1c12c0e1b36aeeece5023f39f9
wrapt==1.14.1
wsproto==1.2.0
xmltodict==0.13.0
xxhash==3.2.0
yarg==0.1.9
yarl==1.8.1
youtube-dl==2021.12.17
youtube-search-python==1.6.6
yt-dlp @ git+https://github.com/yt-dlp/yt-dlp@95a383be1b6fb00c92ee3fb091732c4f6009acb6
ytdl==1.0.0rc5
zeroconf==0.39.1
zipp==3.11.0
zope.interface==5.4.0
